automl-app/backend/requirements.txt
Flask
flask-cors
pandas
numpy
scikit-learn
xgboost
psutil
python-dotenv